# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sneak  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/order-dealer/1": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtiene orders del dealer para repartir",
                "description": "Obtiene las ordenes a repartir de dealer",
                "operationId": "getOrdersByIdDealer",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/order-customer/1": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtiene orders del customer sin entregar",
                "description": "Obtiene las ordenes del customer que no se han repartido todavía",
                "operationId": "getOrdersByIdCustomer",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/order-customer-all/1": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Obtiene totas las orders del customer",
                "description": "Obtiene todas las ordenes del customer",
                "operationId": "getAllOrdersByIdCustomer",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrderResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/order": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Crear nueva orden",
                "description": "Returns order data",
                "operationId": "createdOrder",
                "requestBody": {
                    "description": "Ponga sus credenciales",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address",
                                    "user_id",
                                    "quantity",
                                    "price",
                                    "product_id"
                                ],
                                "properties": {
                                    "address": {
                                        "type": "string",
                                        "example": "C/ Correos"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": "15"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": "99"
                                    },
                                    "product_id": {
                                        "type": "integer",
                                        "example": "5"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario no autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/order/1": {
            "put": {
                "tags": [
                    "Orders"
                ],
                "summary": "Editar orden",
                "description": "Returns order data",
                "operationId": "updateOrder",
                "requestBody": {
                    "description": "Ponga sus credenciales",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "state_id"
                                ],
                                "properties": {
                                    "state_id": {
                                        "type": "integer",
                                        "example": "2"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario no autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "OrderResource": {
                "title": "OrderResource",
                "description": "Order resource",
                "properties": {
                    "data": {
                        "title": "data",
                        "description": "Datos a devolver de la tabla orders",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Order"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "OrderResource"
                }
            },
...
            "Order": {
                "required": [
                    "id, state_id, state, address, created_at, updated_at, userName, productName, OrderLine"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "state_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "state": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "En espera"
                    },
                    "address": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "C/ Correos"
                    },
                    "created_at": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "11/11/11"
                    },
                    "updated_at": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "12/12/12"
                    },
                    "userName": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "Víctor"
                    },
                    "productName": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "Zapatillas Nike"
                    },
                    "OrderLine": {
                        "type": "object",
                        "readOnly": "true",
                        "example": "{ 'id':1, 'quantity':10, 'price':80, 'discount':0, 'order_id':1, 'user_id':1, 'product_id':7 }"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Order"
                }
            },
...
        },
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}